requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
transformers
python-multipart
tf-keras
